feat: Add fadd and fsub

+++ src/megumin/cost/CorrectnessCost.cpp
@@ -37,14 +37,14 @@ int eval_distance(const arm::bits &v1, const arm::bits &v2) {
     return count_set_bits(a ^ b);
 }
 
-uint64_t ulp_distance(const arm::bits& v1, const arm::bits& v2) {
+double ulp_distance(const arm::bits& v1, const arm::bits& v2) {
     int64_t a = v1.data0;
     int64_t b = v2.data0;
 
     a = a < 0 ? std::numeric_limits<int64_t>::min() - a : a;
     b = b < 0 ? std::numeric_limits<int64_t>::min() - b : b;
 
-    uint64_t ulp = a > b ? a - b : b - a;
+    double ulp = a > b ? (double) a - b : (double) b - a;
     // cout << "ulp: " << ulp << endl;
 //    if (ulp > 0) {
 //        assert(false);
@@ -78,9 +78,10 @@ namespace megumin {
         for (int i = 0; i < 32; i++) {
             const bits& reg_target = target_state.fp.get_ref(i, true);
             const bits& reg_rewrite = rewrite_state.fp.get_ref(i, true);
-            result += (double) ulp_distance(reg_target, reg_rewrite);
+            result += ulp_distance(reg_target, reg_rewrite);
         }
 
+        // cout << "result: " << result << endl;
         return result;
     }
